working on decompression

+++ src/ClothoDB/TimeSeriesPoint.h
@@ -12,7 +12,7 @@ public:
     ~TimeSeriesPoint() {};
 
     int32_t timestamp_;
-    int64_t value_;
+    double value_;
 };
 
 }}//incolun::clothodb
